Перевод названий компонентов из kebab-case в camelCase
