+++ src/components/Render/config/filters/standart.js
