+++ src/components/Render/config/index.js
