+++ src/components/Render/helpers/generateImages.js
