+++ src/index.js
@@ -1,10 +1,10 @@
 import "@src/index.css";
 import PageOptions from "@components/PageOptions.js";
 import PreviewOptions from "@components/PreviewOptions.js";
-import RenderOptions from "@components/render";
-import Manuscript from "@components/P5_context2d";
-import Scene from "@components/P5_webGL";
-import FiltersPreview from "@components/FiltersPreview";
+import RenderOptions from "@components/Render";
+import Manuscript from "@components/P5TextList";
+import Scene from "@components/P5NoteScene";
+import FilterWindow from "@components/FilterWindow";
 
 const app = document.querySelector(".app");
 const main = document.createElement("main");
@@ -22,7 +22,7 @@ main.innerHTML = /*html*/ `
 // Добавляем скетчи
 main.querySelector(".p5-sketch-container").appendChild(Manuscript());
 main.querySelector(".p5-sketch-container").appendChild(Scene());
-main.querySelector(".p5-sketch-container").appendChild(FiltersPreview());
+main.querySelector(".p5-sketch-container").appendChild(FilterWindow());
 app.appendChild(main);
 
 main.querySelector(".app__options").appendChild(PageOptions());
